Moved log classes to another folder and separate files
