+++ GentlemanParseDice-DiscordBot/Logger/LogTypeEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace GentelmanParserDiscordBot.Logger
+{
+    public enum LogType
+    {
+        Failed,
+        Roll,
+        Command
+    }
+}
